+++ packages/react/src/experimental/Navigation/Sidebar/Menu/index.stories.tsx
@@ -25,7 +25,7 @@ export const Default: Story = {
   args: {
     tree: [
       {
-        id: 1,
+        id: "1",
         title: "Root",
         items: [
           { label: "Home", icon: Icons.Home, href: "/", exactMatch: true },
@@ -45,7 +45,7 @@ export const Default: Story = {
         isSortable: false,
       },
       {
-        id: 2,
+        id: "2",
         title: "You",
         items: [
           { label: "Me", icon: Icons.Person, href: "/me" },
@@ -56,7 +56,7 @@ export const Default: Story = {
         isSortable: true,
       },
       {
-        id: 3,
+        id: "3",
         title: "Your company",
         items: [
           { label: "Organization", icon: Icons.People, href: "/organization" },